TESTS: new case to test ad_gpo_parse_ini_file()
