+++ setup.cfg
@@ -4,3 +4,4 @@ max-line-length = 119
 [flake8]
 max-line-length = 119
 ignore = W503
+exclude=src/tests/cmocka/GPT.INI
