+++ src/providers/ad/ad_gpo_child.c
@@ -25,10 +25,8 @@
 #include <sys/types.h>
 #include <ctype.h>
 #include <unistd.h>
-#include <sys/stat.h>
 #include <popt.h>
 #include <libsmbclient.h>
-#include <ini_configobj.h>
 #include <security/pam_modules.h>
 
 #include "util/util.h"
@@ -40,8 +38,8 @@
 
 #define SMB_BUFFER_SIZE 65536
 #define GPT_INI "/GPT.INI"
-#define INI_GENERAL_SECTION "General"
-#define GPT_INI_VERSION "Version"
+
+errno_t ad_gpo_parse_ini_file(const char *smb_path, int *_gpt_version);
 
 struct input_buffer {
     int cached_gpt_version;
@@ -269,48 +267,6 @@ done:
     return ret;
 }
 
-static errno_t gpo_sanitize_buffer_content(uint8_t *buf, int buflen)
-{
-    int i;
-    int line_start = 0;
-    int equal_pos = 0;
-
-    if (!buf) {
-        return EINVAL;
-    }
-
-    for (i = 0; i < buflen; ++i) {
-        if (buf[i] == '\n') {
-            line_start = i + 1;
-            continue;
-        }
-        if (buf[i] == '=') {
-            equal_pos = i;
-            continue;
-        }
-        if (isascii(buf[i])) {
-            continue;
-        }
-
-        /* non-ascii */
-        if (equal_pos <= line_start) { /* key */
-            DEBUG(SSSDBG_OP_FAILURE,
-                  "Key or section starting at position %d ('%.*s...') contains"
-                  " non-ascii symbol. File is unusable!\n",
-                  line_start, i - line_start, buf + line_start);
-            return EINVAL;
-        }
-
-        buf[i] = '?';
-        DEBUG(SSSDBG_IMPORTANT_INFO,
-              "Value for key '%.*s' contains non-ascii symbol."
-              " Replacing with '?'\n",
-              equal_pos - line_start, buf + line_start);
-    }
-
-    return EOK;
-}
-
 /*
  * This function stores the input buf to a local file, whose file path
  * is constructed by concatenating:
@@ -458,182 +414,6 @@ gpo_cache_remove_file(const char *smb_path,
 
 done:
     talloc_free(filename);
-    return ret;
-}
-
-static errno_t
-parse_ini_file_with_libini(struct ini_cfgobj *ini_config,
-                           int *_gpt_version)
-{
-    int ret;
-    struct value_obj *vobj = NULL;
-    int gpt_version;
-
-    ret = ini_get_config_valueobj(INI_GENERAL_SECTION, GPT_INI_VERSION,
-                                  ini_config, INI_GET_FIRST_VALUE, &vobj);
-    if (ret != 0) {
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "ini_get_config_valueobj failed [%d][%s]\n", ret, strerror(ret));
-        goto done;
-    }
-    if (vobj == NULL) {
-        DEBUG(SSSDBG_CRIT_FAILURE, "section/name not found: [%s][%s]\n",
-              INI_GENERAL_SECTION, GPT_INI_VERSION);
-        ret = EINVAL;
-        goto done;
-    }
-
-    gpt_version = ini_get_int32_config_value(vobj, 0, -1, &ret);
-    if (ret != 0) {
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "ini_get_int32_config_value failed [%d][%s]\n",
-              ret, strerror(ret));
-        goto done;
-    }
-
-    *_gpt_version = gpt_version;
-
-    ret = EOK;
-
- done:
-
-    return ret;
-}
-
-/*
- * This function parses the GPT_INI file stored in the gpo_cache, and uses the
- * results to populate the output parameters ...
- */
-static errno_t
-ad_gpo_parse_ini_file(const char *smb_path,
-                      int *_gpt_version)
-{
-    struct ini_cfgfile *file_ctx = NULL;
-    struct ini_cfgobj *ini_config = NULL;
-    const char *ini_filename;
-    int ret;
-    int gpt_version = -1;
-    TALLOC_CTX *tmp_ctx = NULL;
-    struct stat st;
-    int fd = -1;
-    uint8_t *buf = NULL;
-
-    tmp_ctx = talloc_new(NULL);
-    if (tmp_ctx == NULL) {
-        ret = ENOMEM;
-        goto done;
-    }
-
-    ini_filename = talloc_asprintf(tmp_ctx, GPO_CACHE_PATH"%s%s",
-                                   smb_path, GPT_INI);
-    if (ini_filename == NULL) {
-        DEBUG(SSSDBG_CRIT_FAILURE, "talloc_asprintf failed.\n");
-        ret = ENOMEM;
-        goto done;
-    }
-
-    DEBUG(SSSDBG_TRACE_FUNC, "ini_filename:%s\n", ini_filename);
-
-    ret = ini_config_create(&ini_config);
-    if (ret != 0) {
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "ini_config_create failed [%d][%s]\n", ret, strerror(ret));
-        goto done;
-    }
-
-    fd = open(ini_filename, O_RDONLY);
-    if (fd == -1) {
-        ret = errno;
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "open() failed [%d][%s]\n", ret, strerror(ret));
-        ret = EIO;
-        goto done;
-    }
-    ret = fstat(fd, &st);
-    if (ret != 0) {
-        ret = errno;
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "stat() failed [%d][%s]\n", ret, strerror(ret));
-        ret = EIO;
-        goto done;
-    }
-    buf = talloc_size(tmp_ctx, st.st_size);
-    if (buf == NULL) {
-        ret = ENOMEM;
-        goto done;
-    }
-    if (sss_atomic_read_s(fd, buf, st.st_size) != st.st_size) {
-        ret = EIO;
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "sss_atomic_read_s() failed\n");
-        goto done;
-    }
-
-    /* Windows uses ANSI (extended-ASCII) to encode the GPT.INI file.
-     * Practically this might mean any code page, including uncompatible
-     * with UTF. Since the only value read by SSSD from GPT.INI is
-     * 'Version=...', just get rid of any non-ascii characters to make
-     * content compatible with lib_iniconfig.
-     */
-    ret = gpo_sanitize_buffer_content(buf, st.st_size);
-    if (ret != EOK) {
-        DEBUG(SSSDBG_CRIT_FAILURE, "gpo_sanitize_buffer_content() failed\n");
-        goto done;
-    }
-
-    ret = ini_config_file_from_mem(buf, st.st_size, &file_ctx);
-    if (ret != 0) {
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "ini_config_file_from_mem() failed [%d][%s]\n", ret, strerror(ret));
-        goto done;
-    }
-
-    ret = ini_config_parse(file_ctx, INI_STOP_ON_NONE, 0, 0, ini_config);
-    if (ret != 0) {
-        int lret;
-        char **errors;
-
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "[%s]: ini_config_parse failed [%d][%s]\n",
-              ini_filename, ret, strerror(ret));
-
-        /* Now get specific errors if there are any */
-        lret = ini_config_get_errors(ini_config, &errors);
-        if (lret != 0) {
-            DEBUG(SSSDBG_CRIT_FAILURE,
-                  "Failed to get specific parse error [%d][%s]\n", lret,
-                  strerror(lret));
-            goto done;
-        }
-
-        for (int i = 0; errors[i]; i++) {
-             DEBUG(SSSDBG_CRIT_FAILURE, "%s\n", errors[i]);
-        }
-        ini_config_free_errors(errors);
-
-        goto done;
-    }
-
-    ret = parse_ini_file_with_libini(ini_config, &gpt_version);
-    if (ret != 0) {
-        DEBUG(SSSDBG_CRIT_FAILURE,
-              "parse_ini_file_with_libini failed [%d][%s]\n",
-              ret, strerror(ret));
-        goto done;
-    }
-
-    *_gpt_version = gpt_version;
-
- done:
-
-    if (ret != EOK) {
-        DEBUG(SSSDBG_CRIT_FAILURE, "Error encountered: %d.\n", ret);
-    }
-
-    ini_config_file_destroy(file_ctx);
-    ini_config_destroy(ini_config);
-    if (fd != -1) close(fd);
-    talloc_free(tmp_ctx);
     return ret;
 }
