+++ src/tests/cmocka/GPT.INI
@@ -1,0 +1,3 @@
+[General]
+Version=6
+displayName=新しいグループ ポリシー オブジェクト
